+++ app/src/main/java/com/example/qoolquotes/ui/screens/BrowseImagesScreen.kt
@@ -59,7 +59,7 @@ fun BrowseImagesScreen(viewModel: BrowseImagesViewModel = hiltViewModel()) {
         ) {
             items(quotes) { quote ->
                 QuoteCard(quote = quote, onClick = {
-                    navController.navigate(QuoteScreenDestination(quote.id.toLong()))
+                    navController.navigate(QuoteScreenDestination(quote.id))
                 })
             }
         }
